+++ packages/diff/diff.ts
@@ -1,5 +1,6 @@
 import { DiffChunk } from '../shared-types'
 import { distance } from 'fastest-levenshtein'
+import { getDiffType, getValueType } from './utils'
 
 type DiffableType =
   | string
@@ -50,37 +51,11 @@ const diffFunc = <T extends DiffableType>(
   }
 }
 
-type ValueOf<T> = T[keyof T]
-type DiffType = ValueOf<Pick<DiffChunk, 'type'>>
-const getDiffType = (diffPower: number): DiffType => {
-  if (diffPower > 0) {
-    return 'add'
-  }
-  if (diffPower < 0) {
-    return 'remove'
-  }
-  return 'change'
-}
-
-const getValueType = (value: unknown): ValueOf<Pick<DiffChunk, 'valueType'>> => {
-  switch (typeof value) {
-    case 'string':
-      return 'string'
-    case 'number':
-      return 'number'
-    case 'boolean':
-      return 'boolean'
-    case 'object':
-      if (Array.isArray(value)) {
-        return 'array'
-      }
-      return value === null ? 'unknown' : 'object'
-    default:
-      return 'unknown'
-  }
-}
-
-const diffImpl = <T extends DiffableType>(path: string, nextState: T, prevState: T): DiffChunk => {
+export const diffImpl = <T extends DiffableType>(
+  path: string,
+  nextState: T,
+  prevState: T
+): DiffChunk => {
   const { diff, diffPower } = diffFunc(prevState, nextState)
 
   const diffChunk: DiffChunk = {
@@ -94,9 +69,6 @@ const diffImpl = <T extends DiffableType>(path: string, nextState: T, prevState:
 
   return diffChunk
 }
-
-const diff = diffImpl
-export default diff
 
 export const isDiffable = (value: unknown): value is DiffableType => {
   return (
