+++ packages/diff/index.ts
@@ -1,1 +1,6 @@
-export * from './diff'
+import { diffImpl, isDiffable } from './diff'
+
+const diff = diffImpl
+
+export default diff
+export { isDiffable }
