+++ packages/sonification/index.ts
@@ -1,5 +1,1 @@
-export { sonifyChanges, playSonicChunk } from './sonification'
-
-export { AudioContextManager } from './utils'
-
-export { AUDIO_CONFIG } from './constants'
+export { sonifyChanges } from './sonification'
