--- packages/sonification/types.d.ts
@@ -1,1 +1,0 @@
-export { SonicChunk } from '../shared-types'
