+++ packages/sonification/utils.ts
@@ -162,14 +162,3 @@ export class AudioContextManager {
     // No dialog cleanup needed here
   }
 }
-
-// --- simpleHash remains the same ---
-/** Generate a simple hash from a string */
-export const simpleHash = (str: string): number => {
-  let hash = 0
-  for (let i = 0; i < str.length; i++) {
-    hash = (hash << 5) - hash + str.charCodeAt(i)
-    hash |= 0 // Convert to 32bit integer
-  }
-  return Math.abs(hash)
-}
